requirements/base.txt
Django==2.0.3
django-filter==1.1.0
djangorestframework==3.7.7
djangorestframework-jwt==1.11.0
environ==1.0
Pillow==5.0.0
PyJWT==1.5.3
pytz==2018.3
